Implement AnyWidget communication protocol & cleanup
This replaces backend<-->frontend communications with the widget with a
custom Comm object. This is the protocol that Jupyter uses to
communicate with widgets, so it should provide the most compatibility
with widgets.

Something to note is that the widget is not rendered until the initial
"update" message is received. This means about a second or two of delay
before it renders. I've added a placeholder message and a message in
case there's an error while loading the plugin (E.g. invalid javascript
syntax).

I also had to pull in `lodash` via npm in order to do deep equals, and
added calls to the cleanup functions.

This has some code borrowed from other implementations, so I've included
their licenses.